+++ R/add_sudo.R
@@ -1,0 +1,4 @@
+add_sudo <- function(sudo,
+                     pass = Sys.getenv("PASS")){
+  if(sudo) paste("echo",pass,"|","sudo -S") else ""
+}