+++ R/messager.R
@@ -12,7 +12,9 @@
 #' 
 #' @return Null 
 #' @keywords internal 
-messager <- function(..., v = TRUE, parallel = FALSE) {
+messager <- function(..., 
+                     v = Sys.getenv("VERBOSE")!="FALSE", 
+                     parallel = FALSE) {
     if(parallel){
         if(v) try({message_parallel(...)})
     } else {