+++ tests/testthat/test-get_os.R
@@ -1,0 +1,5 @@
+test_that("get_os works", {
+  
+  os <- get_os()
+  testthat::expect_true(tolower(os) %in% c("windows","mac","linux"))
+})