+++ tests/testthat/testfun.R
@@ -1,0 +1,7 @@
+testfun <- function(x=1,y=2){
+  print("Running testfun")
+}
+
+testfun2 <- function(x2=3,y2=4){
+  print("Running testfun2")
+}