+++ NEWS.md
@@ -1,3 +1,19 @@
+# devoptera 0.99.1
+
+## New features
+
+* Update *rworkflows.yml*
+* Add Docker vignette
+* `args2vars`: 
+  - add `set_dots` arg.
+  - Enable inference of `fn` arg from currently open file (default), 
+    R source file path, parent function, or explicit function.
+* New exported functions:
+  - `current_file`
+  - `list_functions`
+  - `stop_early`
+* Improve code coverage.
+
 # devoptera 0.99.0
 
 * Added a `NEWS.md` file to track changes to the package.